backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
supabase==2.0.0
httpx==0.24.1
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
